+++ packages/panel/src/router/index.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import VueRouter, {RouteConfig} from 'vue-router';
-import Home from '../views/Home.vue';
+import Home from '../views/Index.vue';
 import Register from '../views/auth/Register.vue';
 import Login from '../views/auth/Login.vue';
